src/requirements.txt
django==3.1.7
celery==5.0.5
gunicorn==20.0.4
djangorestframework==3.12.2
markdown==3.3.4
django-filter==2.4.0
kafka-python==2.0.2
python-snappy==0.6.0
crc32c==2.2
confluent-kafka==1.6.0
locket==0.2.1
